# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Wallet",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/wallet": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "parameters": [
                    {
                        "description": "operation info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateWalletDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/wallets/{WALLET_UUID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID wallet",
                        "name": "WALLET_UUID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "dto.UpdateWalletDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "operationType": {
                    "type": "string",
                    "enum": [
                        "DEPOSIT",
                        "WITHDRAW"
                    ]
                },
                "valletId": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        }
    }
}